app/requirements.txt
Django==1.3.1
Fabric==1.3.3
MySQL-python==1.2.3
Pygments==1.4
South==0.7.3
Unidecode==0.04.9
Unipath==0.2.1
Whoosh==2.3.0
bpython==0.10.1
django-appconf==0.4.1
django-compressor==1.0.1
django-debug-toolbar==0.8.5
django-haystack==1.2.5
django-pagination==1.0.7
django-taggit==0.9.3
django-templatetag-sugar==0.1
easy-thumbnails==1.0-alpha-18
html5lib==0.95
pisa==3.0.33
pycrypto==2.4.1
reportlab==2.5
requests==0.8.3
ssh==1.7.11
wsgiref==0.1.2
xlrd==0.7.7
